Time: 1 ms (18.83%), Space: 13.8 MB (5.07%) - LeetHub

+++ 0274-h-index/0274-h-index.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int hIndex(vector<int>& citations) {
-        int n = citations.size();
+        short int n = citations.size();
         std :: sort(citations.rbegin() , citations.rend());
 
         for(int i {0} ; i < n ; i++){
